Ajout de donnée au seeder et mise en place de la platforme en ligne

+++ api/ExerciceAPI.js
@@ -30,24 +30,56 @@ module.exports = (app, ExerciceService, jwt) => {
     });
 
     app.post("/exercice/favori", jwt.validateJWT, (req, res) => {
-        const exercice_id = req.body['exercice_id']
-        // Ajout de l'exercice en favori pour le User
-        ExerciceService.dao.insertExoFavori(exercice_id, req.user.id)
-            .then(_ => res.status(200).end())
-            .catch(e => {
-                console.log(e)
-                res.status(500).end()
+        const exercice_id = req.body['exercice_id'];
+        if (exercice_id == null || exercice_id == undefined) {
+            return res.status(404).json({ error: "Exercice not found" }).end();
+        }
+
+        // Vérifiez si l'exercice_id existe dans la base de données
+        ExerciceService.dao.getById(exercice_id)
+            .then(exercice => {
+                if (!exercice) {
+                    return res.status(404).json({ error: "Exercice not found" }).end();
+                }
+
+                // Ajout de l'exercice en favori pour le User
+                ExerciceService.dao.insertExoFavori(exercice_id, req.user.id)
+                    .then(_ => res.status(200).end())
+                    .catch(e => {
+                        console.error(e);
+                        res.status(500).end();
+                    });
             })
-    })
+            .catch(e => {
+                console.error(e);
+                res.status(500).end();
+            });
+    });
 
     app.delete("/exercice/favori", jwt.validateJWT, (req, res) => {
-        const exercice_id = req.body['exercice_id']
-        // Delete de l'exercice en favori pour le User
-        ExerciceService.dao.deleteExoFavori(exercice_id, req.user.id)
-            .then(_ => res.status(200).end())
-            .catch(e => {
-                console.log(e)
-                res.status(500).end()
+        const exercice_id = req.body['exercice_id'];
+        if (exercice_id == null || exercice_id == undefined) {
+            return res.status(404).json({ error: "Exercice not found" }).end();
+        }
+
+        // Vérifiez si l'exercice_id existe dans la base de données des favoris du User
+        ExerciceService.dao.getByIdFavori(req.user.id, exercice_id)
+            .then(exercice => {
+                if (!exercice || (Array.isArray(exercice) && exercice.length === 0)) {
+                    return res.status(404).json({ error: "Exercice not found" }).end();
+                }
+
+                // Suppression de l'exercice en favori pour le User
+                ExerciceService.dao.deleteExoFavori(exercice_id, req.user.id)
+                    .then(_ => res.status(200).end())
+                    .catch(e => {
+                        console.error(e);
+                        res.status(500).end();
+                    });
             })
-    })
+            .catch(e => {
+                console.error(e);
+                res.status(500).end();
+            });
+    });
 }
